refactor: dont use a worker for version check
This means we don't need `--unstable` for the installation anymore.
