+++ README.md
@@ -5,7 +5,7 @@ Command line tool for Deno Deploy.
 ## Install
 
 ```shell
-deno install --allow-read --allow-write --allow-env --allow-net --allow-run --no-check --unstable -r -f https://deno.land/x/deploy/deployctl.ts
+deno install --allow-read --allow-write --allow-env --allow-net --allow-run --no-check -r -f https://deno.land/x/deploy/deployctl.ts
 ```
 
 ## Usage
