+++ deps.ts
@@ -6,14 +6,14 @@ export {
   join,
   resolve,
   toFileUrl,
-} from "https://deno.land/std@0.96.0/path/mod.ts";
+} from "https://deno.land/std@0.97.0/path/mod.ts";
 export {
   bold,
   green,
   red,
   yellow,
-} from "https://deno.land/std@0.96.0/fmt/colors.ts";
-export { parse as parseArgs } from "https://deno.land/std@0.96.0/flags/mod.ts";
+} from "https://deno.land/std@0.97.0/fmt/colors.ts";
+export { parse as parseArgs } from "https://deno.land/std@0.97.0/flags/mod.ts";
 
 // x/semver
 export {
@@ -22,7 +22,7 @@ export {
 } from "https://deno.land/x/semver@v1.4.0/mod.ts";
 
 // x/cache
-export { cache } from "https://deno.land/x/cache@0.2.12/mod.ts";
+export { cache } from "https://deno.land/x/cache@0.2.13/mod.ts";
 
 // x/dotenv
 export { config as dotEnvConfig } from "https://deno.land/x/dotenv@v2.0.0/mod.ts";
