+++ src/version.ts
@@ -2,4 +2,4 @@ export const VERSION = "0.3.0";
 
 // The minium Deno version required. Currently 1.9.1 because we are making use of
 // the --unstable `Deno.startHttp` API, and 1.9.0 still has some bugs.
-export const MINIMUM_DENO_VERSION = "1.9.1";
+export const MINIMUM_DENO_VERSION = "1.11.1";
